Move dashboard display story into own js module
Also update the import path after moving the dashboard.
