+++ ng/src/web/stories/dashboard/index.js
@@ -20,55 +20,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
  */
-import React from 'react';
 
-import {Div} from 'glamorous';
-
-import {storiesOf} from '@storybook/react';
-import {action} from '@storybook/addon-actions';
-
-import MenuEntry from 'web/components/menu/menuentry';
-
-import Display from 'web/components/dashboard/display';
-import DisplayMenu from 'web/components/dashboard/displaymenu';
-
-const removeaction = action('on remove click');
-
-storiesOf('Dashboard/Display', module)
-  .add('default', () => {
-    return (
-      <Div width="500px">
-        <Display
-          title="Foo Bar"
-          onRemoveClick={removeaction}>
-          <Div
-            backgroundColor="blue"
-            width="400px"
-            height="400px"
-          />
-        </Display>
-      </Div>
-    );
-  })
-  .add('with menu', () => (
-    <Div width="500px">
-      <Display
-        title="Foo Bar"
-        onRemoveClick={removeaction}
-        menu={
-          <DisplayMenu>
-            <MenuEntry>Download Stuff</MenuEntry>
-            <MenuEntry>Do Something</MenuEntry>
-          </DisplayMenu>
-        }
-      >
-        <Div
-          backgroundColor="blue"
-          width="400px"
-          height="400px"
-        />
-      </Display>
-    </Div>
-  ));
+import './display';
 
 // vim: set ts=2 sw=2 tw=80:
